Additional changes in the configuration and minor fixes.
