+++ karyon-examples/hello-netflix-oss/src/main/webapp/WEB-INF/web.xml
@@ -1,9 +1,7 @@
 <?xml version="1.0"?>
-<web-app xmlns="http://java.sun.com/xml/ns/javaee"
+<web-app version="2.5" xmlns="http://java.sun.com/xml/ns/javaee"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-	      http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-         version="3.0">
+         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee">
 
     <filter>
         <filter-name>guiceFilter</filter-name>
