+++ karyon-examples/hello-netflix-oss/src/test/resources/arquillian.xml
@@ -4,4 +4,12 @@
             xsi:schemaLocation="http://jboss.org/schema/arquillian
     http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
 
+    <engine>
+        <property name="deploymentExportPath">build/deployments</property>
+    </engine>
+
+    <extension qualifier="karyon">
+        <property name="autoPackage">true</property>
+    </extension>
+
 </arquillian>
