+++ karyon-extensions-testsuite/src/main/java/com/netflix/karyon/server/test/configuration/ConfigurationParser.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2013 Netflix, Inc.
+ *
+ *      Licensed under the Apache License, Version 2.0 (the "License");
+ *      you may not use this file except in compliance with the License.
+ *      You may obtain a copy of the License at
+ *
+ *          http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *      Unless required by applicable law or agreed to in writing, software
+ *      distributed under the License is distributed on an "AS IS" BASIS,
+ *      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *      See the License for the specific language governing permissions and
+ *      limitations under the License.
+ */
+package com.netflix.karyon.server.test.configuration;
+
+/**
+ * Parses the configuration. The implementation classes can support different storage mechanisms and data
+ * representations.
+ *
+ * @author Jakub Narloch (jmnarloch@gmail.com)
+ */
+public interface ConfigurationParser {
+
+    /**
+     * Parses the configuration.
+     *
+     * @return the configuration
+     */
+    KaryonExtensionConfiguration parse();
+}
