+++ karyon-extensions-testsuite/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-com.netflix.kayron.server.test.client.KayronExtension
+com.netflix.karyon.server.test.client.KaryonExtension
