+++ karyon-extensions-testsuite/src/test/resources/arquillian.xml
@@ -5,8 +5,9 @@
         http://jboss.org/schema/arquillian
         http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
 
-    <extension qualifier="kayron">
+    <extension qualifier="karyon">
         <property name="applicationId">applicationId</property>
         <property name="environment">environment</property>
+        <property name="autoPackage">true</property>
     </extension>
 </arquillian>
